+++ docs/changelog.rst
@@ -4,6 +4,9 @@ Changelog
 `Unreleased`_ - TBD
 -------------------
 
+`2.8.3`_ - 2020-11-27
+---------------------
+
 **Added**
 
 - Display failed response payload in the error output for the ``pytest`` plugin. `#895`_
@@ -1448,7 +1451,8 @@ Deprecated
 
 - Initial public release
 
-.. _Unreleased: https://github.com/schemathesis/schemathesis/compare/v2.8.2...HEAD
+.. _Unreleased: https://github.com/schemathesis/schemathesis/compare/v2.8.3...HEAD
+.. _2.8.3: https://github.com/schemathesis/schemathesis/compare/v2.8.2...v2.8.3
 .. _2.8.2: https://github.com/schemathesis/schemathesis/compare/v2.8.1...v2.8.2
 .. _2.8.1: https://github.com/schemathesis/schemathesis/compare/v2.8.0...v2.8.1
 .. _2.8.0: https://github.com/schemathesis/schemathesis/compare/v2.7.7...v2.8.0